safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/script-src-self-blocked-01-expected.txt

CONSOLE MESSAGE: Refused to load the script 'http://localhost:8000/security/contentSecurityPolicy/resources/script.js' because it violates the following Content Security Policy directive: "script-src 'self'".



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
